llamaindex_QA/requirement.txt
llama-index==0.9.44
python-dotenv
html2text